sorta kinda working
We're hacking the OPTree but rather than setting the values
SvREADONLY_on() inside the OPs that are being generated, we need to
inject *new* OPs to set those values to read only *after* the expression
is run.

We also need to figure out a better way to identify the SVs in question.